[OK] 33 files have been changed by Rector

+++ application/modules/layout/views/alerts.php
@@ -4,10 +4,8 @@ $alert_class .= isset($without_margin) ? ' no-margin' : '';
 $content = null;
 
 // Get validation errors
-if (function_exists('validation_errors')) {
-    if (validation_errors()) {
-        echo validation_errors('<div class="' . $alert_class . ' alert-danger">', '</div>');
-    }
+if (function_exists('validation_errors') && validation_errors()) {
+    echo validation_errors('<div class="' . $alert_class . ' alert-danger">', '</div>');
 }
 
 // Get flash alert_*type* messages and show them
